+++ CHANGELOG.md
@@ -10,6 +10,17 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ## \[Unreleased\]
 
+## Added
+
+- `Alert`:
+    - Added the following fields to `Alert.attributes`: `n_previous_sources` and `ssobjectid`.
+- Unit tests:
+    - Tests take into account new `Alert.attributes` keys.
+
+### Changed
+
+- Rename `Alert` method `_add_id_attributes()` -> `_add_attributes()` to reflect that more properties are added.
+
 ### Fixed
 
 - `Topic.publish()`: Message attributes were improperly defined